requirements.txt
PyGithub==1.26.0
slacker==0.9.2
